weather-alert-app/docker/faust/requirements.txt
kafka-python
faust
fastavro
confluent-kafka
requests
avro-python3
twilio
